refactor: 将组价基类重命名为 gui
